requirements.txt

fastapi==0.115.0
uvicorn==0.30.1


streamlit==1.37.1

google-genai==0.3.0  

python-dotenv==1.0.1


pydantic==2.9.1
requests==2.32.3


openpyxl==3.1.5  

google-generativeai